This is the script for multiple files
This is the script for pre-processing multiple classes from Open Images for Training yolo; please read the comments in the script on how to use.  I have not tested this more than once, and the weights it generated were OK at best (I believe the issue with it detecting Cat and Banana poorly was the result of the dataset for Beer  being massively larger than the two; we also of course are ignoring test and validation data too per the original tutorial).  It will be some time before I update the documentation in the README, but this script does work so long as you adjust the config for yolo and other associated files as well.